feat: refactor documentation and remove unused components; add new useFragmentToken hook
