+++ web/src/platform/routes/Docs/Sdk/Storage.tsx
@@ -4,39 +4,9 @@ import { Text } from '@astryxdesign/core/Text';
 import { Stack } from '@astryxdesign/core/Stack';
 import { Heading } from '@astryxdesign/core/Heading';
 import { CodeBlock } from '@astryxdesign/core/CodeBlock';
+import { Table, TableBody, TableCell, TableHeader, TableHeaderCell, TableRow } from '@astryxdesign/core/Table';
 import { CheckCheck, CheckCircle, Wrench } from 'lucide-react';
 import { DocsArticle, createDocsMeta } from '@/platform/routes/Docs/Article';
-import { EnvironmentTable, type EnvironmentRow } from '@/platform/docs/sdk/EnvironmentTable';
-
-const environments: EnvironmentRow[] = [
-    {
-        name: 'Testing',
-        icon: CheckCheck,
-        backend: (
-            <>
-                <Code>memory</Code> backend for isolated in-memory test files.
-            </>
-        ),
-    },
-    {
-        name: 'Development',
-        icon: Wrench,
-        backend: (
-            <>
-                <Code>file</Code> backend for inspectable local files.
-            </>
-        ),
-    },
-    {
-        name: 'Production',
-        icon: CheckCircle,
-        backend: (
-            <>
-                <Code>s3</Code> backend using application and shared prefixes in one Organization bucket.
-            </>
-        ),
-    },
-];
 
 export const metadata = {
     path: '/docs/sdk/storage',
@@ -69,7 +39,54 @@ function Content() {
                 </Link>
                 . Application code uses the same filesystem interface in local development, tests, and production.
             </Text>
-            <EnvironmentTable environments={environments} />
+            <Table>
+                <TableHeader>
+                    <TableRow>
+                        <TableHeaderCell>Environment</TableHeaderCell>
+                    </TableRow>
+                </TableHeader>
+                <TableBody>
+                    <TableRow>
+                        <TableCell>
+                            <Stack gap={1}>
+                                <Stack direction="horizontal" gap={2} align="center">
+                                    <CheckCheck aria-hidden="true" className="text-accent" size={16} />
+                                    <Text weight="semibold">Testing</Text>
+                                </Stack>
+                                <Text type="supporting">
+                                    <Code>memory</Code> backend for isolated in-memory test files.
+                                </Text>
+                            </Stack>
+                        </TableCell>
+                    </TableRow>
+                    <TableRow>
+                        <TableCell>
+                            <Stack gap={1}>
+                                <Stack direction="horizontal" gap={2} align="center">
+                                    <Wrench aria-hidden="true" className="text-accent" size={16} />
+                                    <Text weight="semibold">Development</Text>
+                                </Stack>
+                                <Text type="supporting">
+                                    <Code>file</Code> backend for inspectable local files.
+                                </Text>
+                            </Stack>
+                        </TableCell>
+                    </TableRow>
+                    <TableRow>
+                        <TableCell>
+                            <Stack gap={1}>
+                                <Stack direction="horizontal" gap={2} align="center">
+                                    <CheckCircle aria-hidden="true" className="text-accent" size={16} />
+                                    <Text weight="semibold">Production</Text>
+                                </Stack>
+                                <Text type="supporting">
+                                    <Code>s3</Code> backend using application and shared prefixes in one Organization bucket.
+                                </Text>
+                            </Stack>
+                        </TableCell>
+                    </TableRow>
+                </TableBody>
+            </Table>
             <Heading id="usage" level={2}>
                 Usage
             </Heading>
